player: Synchronize terminal updates with audio
After about 5 minutes, the terminal starts to get visibly behind the
audio. This is because it is scheduled with respect to when setTimeout
fired and the next event -- but setTimeout frequently fires late.

Instead, schedule the next event based on its position with the audio.
This keeps things in sync.

Thanks to @marchelzo for the idea.