player: fix seeking bug
This probably should have been obvious, but whatever. xterm.js buffers
its writes in case it's taking a long time to update the DOM. It
therefore also has the ability to use flow control to tell applications
writing to the pty to stop sending data.

But we're not a pty application, and we're just calling write() in a
loop with relatively short stuff. This ends up causing a setTimeout for
nearly every write, and eventually JavaScript runs out of timer handles.
At this point, no more updates occur.

Instead, we should just buffer all of our input and only write once.
This is now both snappy and correct, even for large videos.