Switch to libsoundio
The API is nicer. It's less portable, but it has better support for
Linux systems (or at least claims to), in particular PulseAudio on
Ubuntu. We're losing support for things like OSS in the BSDs, but I can
live with that.

This also improves handling of recording formats and sample rates, and
removes the dependency on Concurrency Kit, since libsoundio has a
performant, thread-safe SPSC ring buffer.