Makefile: use gcc -M flags to get complete DAG on recompiles
This patch allows the Makefile to see a complete DAG on recompiles so
that, for example, changes to the headerfile castty.h will cause all
objects of source files including that header to get re-compiled.

Obviously on the first run of make this will not be the case, however,
on the first run this doesn't matter.

This patch makes use of some gnu make specific features, however, the
original makefile already made the assumption of gnu make.