audio.c: Fix use of uninitialised "lame" when not using mp3
Fix uninitialised use of variable `lame` in function writer when mp3
output is not enabled. Additionally, lame_close was also being called on
a potentially uninitialised value.

Since libmp3lame handles passing a NULL to any of its functions
gracefully (although this is not documented), the variable `lame` is now
also initialised to NULL, this also prevents future accidents where
calls to libmp3lame functions are not guarded with an if (mp3).

Finally, this also silences a gcc warning about potentially
uninitialised use of the variable `lame`.