castty, player: use asciicast file format
https://github.com/asciinema/asciinema/blob/master/doc/asciicast-v1.md

Asciinema has already defined a file format for this sort of thing
(which I probably would have known if I knew about asciinema before I
started writing this thing). For the sake of interoperability, I'd like
to switch to this format.

This changes a lot of things, but the big two are:

 * Events are now offset from the previous event. This means the player
   must now keep track of the duration played to sync video with audio.
   Also, events are now represented as fractional seconds, so that
   changed things too.

 * UTF-8 is natively supported, but is now output as \uxxxx sequences
   into the JSON. This is much nicer. Thanks to Björn Höhrmann for
   http://bjoern.hoehrmann.de/utf-8/decoder/dfa/.