castty: Don't close stdin of subchild on SIGCHLD (fixes #25)
outputproc already closes stdin before going into its poll loop, this
means that when the SIGCHLD handler closes stdin for subchild, it ends
up closing an already closed fd.

Additionally, on my system using pulseaudio, this was closing the read
end of an internal pulseaudio pipe causing a stacktrace on exit. This
fixes #25.