Makefile: use $(MAKE), de-duplicate recipes and add .PHONY target
This patch uses $(MAKE) for recursion, de-duplicates the recipes of the
root makefile by making use of a .DEFAULT target and adds a .PHONY
target for all rules in the root makefile.

$(MAKE) is a special variable which should be used when make is being
ran recursively.

The recipe of the target named .DEFAULT will execute if a rule for a
target cannot be found. As running make with no target means that
.DEFAULT's recipe will not execute, therefore a single goal has to be
included to be used as a default goal for these scenarios.

The rule `$(MAKE) -C src $@` will pass the current target alongside any
relevant flags to the makefile in the src directory.

This change also allows setting the .DEFAULT_GOAL variable in the
config.mk to control the default goal in the src makefile.

Finally, this change fixes parallel builds by using $(MAKE).