Disambiguate selection of raw devices
The libsoundio docs note that a device ID may be present multiple times
depending on whether there is access to a device both through a proxy
service (like Pulse) and directly to the device itself.

This change adds a -R flag that allows selection of a raw device. It
also disambiguates device IDs that are available through both means when
enumerating devices, and updates the suggested castty string to include
the -R flag in the case that the enumerated device is a raw device.